Add files via upload
For this project, I coded a Java based program that will read from a file and display to the console all the words in the file in alphabetical order along with the number of times each word appeared in the file. Downloading the Project4.java and TextFileInput.java should be fine enough to compile and run the program. You can also use the Project4.jar which has all necessary files for this project zipped. Project4.jar is an executable/runnable jar file, but as this program prints the desired output in the console, you will have to unzip it and run it through an IDE or terminal anyway.

This project is built using TreeMap data structure which implements Red-Black Tree. Hence, the structure of this program has a time complexity of O(log n), similar to TreeList or a balanced Binary Search Tree (BST).